完成博客文章数据库的建立和自动化管理，博客文章罗列，自动化测试

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,4 +1,4 @@
-create table user(id int primary key auto_increment ,
+create table user(id bigint primary key auto_increment ,
   username varchar(10) unique,
   encrypted_password varchar(100),
   avatar varchar(100),
